create a seperate controller for generate pdf in new form
